Move APIv1 converters into a separate legacy header
