+++ src/details/ArborX_IndexableGetter.hpp
@@ -40,43 +40,6 @@ struct DefaultIndexableGetter
   }
 };
 
-template <typename Primitives, typename BoundingVolume>
-class LegacyValues
-{
-  Primitives _primitives;
-  using Access = AccessTraits<Primitives, PrimitivesTag>;
-
-public:
-  using memory_space = typename Access::memory_space;
-  using value_type = Details::PairIndexVolume<BoundingVolume>;
-  using size_type =
-      Kokkos::detected_t<Details::AccessTraitsSizeArchetypeExpression, Access,
-                         Primitives>;
-
-  LegacyValues(Primitives const &primitives)
-      : _primitives(primitives)
-  {}
-
-  KOKKOS_FUNCTION
-  decltype(auto) operator()(size_type i) const
-  {
-    if constexpr (std::is_same_v<BoundingVolume,
-                                 typename AccessTraitsHelper<Access>::type>)
-    {
-      return value_type{(unsigned)i, Access::get(_primitives, i)};
-    }
-    else
-    {
-      BoundingVolume bounding_volume{};
-      expand(bounding_volume, Access::get(_primitives, i));
-      return value_type{(unsigned)i, bounding_volume};
-    }
-  }
-
-  KOKKOS_FUNCTION
-  size_type size() const { return Access::size(_primitives); }
-};
-
 template <typename Primitives>
 struct Indexables
 {
